pyproject.toml
[build-system]
# setuptools-scm is used to dynamically set the version from git tags
requires = ["setuptools>=61.0", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "zone-poker"
dynamic = ["version"] # The version is now managed by setuptools-scm
authors = [
  { name="BranchingBad", email="contact@branchingbad.com" },
]
description = "A professional DNS reconnaissance and OSINT tool for comprehensive domain analysis."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Security",
    "Topic :: Internet :: Name Service (DNS)",
    "Environment :: Console",
]
keywords = ["dns", "osint", "reconnaissance", "security", "domain", "subdomain", "axfr"]
# Project dependencies
dependencies = [
    "dnspython~=2.6",
    "httpx~=0.27",
    "python-whois~=0.8",
    "ipwhois~=1.2",
    "rich~=14.2",
    "beautifulsoup4~=4.12",
    "mmh3~=5.2",
    "PyYAML>=6.0.1,<7.0",
    "tldextract~=5.3"
]

[project.urls]
Homepage = "https://github.com/BranchingBad/zone-poker"
Issues = "https://github.com/BranchingBad/zone-poker/issues"

[project.scripts]
zone-poker = "zone_poker:main_wrapper"

[project.optional-dependencies]
dev = [
    "pytest~=9.0", # ruff replaces flake8 and black
    "pre-commit~=4.4",
    "pytest-cov~=7.0",
    "pytest-asyncio~=1.3",
    "bandit~=1.7",
    "pip-audit~=2.9",
    "respx~=0.22",
    "pip-tools~=7.4",
    "ruff" # Add ruff as a dev dependency
]

[tool.setuptools]
# Explicitly include the top-level script module
py-modules = ["zone_poker"]
package-data = { "modules.analysis" = ["*.json"] }

[tool.setuptools.packages.find]
# Tell setuptools to find the 'modules' package and all its subpackages
include = ["modules*"]
# It's also good practice to explicitly exclude tests
exclude = ["tests*"]

[tool.setuptools_scm]
# This configuration tells setuptools-scm to write the dynamic version
# to a file, which makes it accessible at runtime.
write_to = "modules/_version.py"

[tool.ruff]
# Set the maximum line length.
line-length = 140

[tool.ruff.lint]
# Enable the same rule set as flake8.
select = ["E", "F", "W"]
ignore = [] # Add E402 if you have unavoidable circular imports

[tool.ruff.format]
# Use black-compatible formatting.
quote-style = "double"
